[refactor]: rename utils module class file names snake_case
