+++ spatialetl/utils/bing/__init__.py
@@ -10,6 +10,6 @@ BING (:mod:`coverage.io.netcdf.symphonie`)
 from __future__ import division, print_function, absolute_import
 
 
-from .BingImageRetrieval import BingImageRetrieval
+from .bing_image_retrieval import BingImageRetrieval
 
 __all__ = ['BingImageRetrieval']
