+++ spatialetl/utils/bing/core/aerial_image_retrieval.py
