+++ spatialetl/utils/bing/core/tiff_reader.py
