+++ spatialetl/utils/bing/core/tiff_writer.py
