+++ spatialetl/utils/variable_definition.py
